pyproject.toml
[tool.poetry]
name = "cwb-bronze-layer"
version = "0.1.0"
description = ""
authors = ["DidierArchambault <didier.archambault@augmentisconsulting.com>"]
readme = "README.md"
packages = [{ include = "dal" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pyspark = "3.5.5"
delta-spark = ">=3.2,<3.3"
pytest = ">=9.0.2,<10.0.0"
toml = ">=0.10.2,<0.11.0"
duckdb = ">=1.4.4,<2.0.0"
pandas = ">=2.0,<3.0"  # important: pandas 3.0 n'existe pas aujourd'hui
deltalake = ">=1.4.1,<2.0.0"

[tool.poetry.scripts]
main = "dal.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
